SAS Soil calculations to see if fire triggered
still in development, but a workable chunk of code.  First, you have to calculate the soil moisture dynamics for the site and then figure out how often soil is likely to be tiggered in each of our SAS age groups.  Note: Disturbance (fire) is a MONTHLY thing, so need to look at how many months/yr it's likely to be triggered.