Fixed a Bug
One of the years had an issue with stack() because it had only 1 cohort, so R had built the data frame to already be vertical. This isn't the most robust fix, but it's the only way I know how to do it and I've tested it. It works, but it might take slightly longer to compute.